+++ src/main/java/br/com/rogerio/Expression.java
@@ -3,4 +3,5 @@ package br.com.rogerio;
 interface Expression {
 Money reduce(Bank bank, String to);
 Expression plus(Expression addend);
+Expression times(int multipler);
 }