+++ src/main/java/br/com/rogerio/Sum.java
@@ -18,7 +18,12 @@ public class Sum implements Expression {
 
 	@Override
 	public Expression plus(Expression addend) {
-		return null;
+		return new Sum(this, addend);
+	}
+
+	@Override
+	public Expression times(int multipler) {
+		return new Sum(augend.times(multipler), addend.times(multipler));
 	}
 
 }